Expts with multiplicative errors + multiplicative/additive bias terms
A bit of tidying up after the reset back to version 1 VSEM
Added new experiments with multiplicative obs errors

Under "addressing errors"
Added a multiplicative obs bias term to the likelihood which seems to
find the error.
Also added multiplicative and additive bias terms for the unbalanced
model error experiment which improved the fit for the low data output
(Cv).